dz10 | headers and ChatInfo
