+++ react-chat/src/components/ChatInfo/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatInfo.jsx'
