+++ react-chat/src/page/PageChatInfo/PageChatInfo.module.scss
@@ -4,115 +4,4 @@
   align-items: center;
   padding: 10px;
   height: calc(100vh - 70px);
-
-  .title {
-    font-size: 18px;
-    text-align: start;
-  }
-}
-
-.inner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  gap: 10px;
-  max-width: 500px;
-
-  .title {
-    font-size: 16px;
-  }
-}
-
-.avatar {
-  height: 50px;
-  width: 50px;
-  object-fit: cover;
-  border-radius: 50%;
-  box-shadow: rgba(99, 99, 99, 0.4) 0px 2px 8px 2px;
-}
-
-.head {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 15px;
-  width: 100%;
-  padding: 0 15px;
-  margin-bottom: 15px;
-
-  .avatar {
-    width: 80px;
-    height: 80px;
-  }
-
-  .title {
-    font-size: 20px;
-    -webkit-line-clamp: 3;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-  }
-}
-
-.membersList {
-  font-size: 14px;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 15px;
-  width: 100%;
-  background-color: var(--third-white-color);
-  overflow-y: auto;
-  scrollbar-width: thin;
-  scrollbar-color: var(--transparent-dark-green) var(--second-white-color);
-  border-radius: 15px;
-
-  .membersInner {
-    margin: 0;
-    min-height: 100%;
-    display: flex;
-    flex-direction: column;
-    overflow-x: hidden;
-    gap: 15px;
-    width: 100%;
-    padding: 15px;
-  }
-}
-
-.leave {
-  width: 200px;
-  min-height: 40px;
-  border-radius: 25px;
-  cursor: pointer;
-  transition: all .1s;
-  color: white;
-  background-color: red;
-  transition: all .1s ease;
-
-  &:hover {
-    scale: 1.05;
-  }
-
-  &:active {
-    scale: 0.95;
-    background-color: darkred;
-  }
-}
-
-@media (min-width: 900px) {
-  .head {
-    .avatar {
-      width: 120px;
-      height: 120px;
-    }
-
-    .title {
-      font-size: 24px;
-    }
-  }
-
-  .inner .title {
-    font-size: 20px;
-  }
 }
